moved over to using haskell
In for a penny, in for a pound. Move over to using Haskell for an
implementation language. This way I get to learn all about a new
language, a new paradigm and I get to try and make Haskell talk nicely
to AWS. If this works, I'll have learned a _lot_.